Add shared ninja jobserver support
This patch depends on the patch "Add shared make jobserver support"
for creating a named pipe with execution tokens.
This allows both make and ninja to share the same pool of tokens allowing
the total number of compile tasks to be restricted to the number of threads.
On kirkstone cmake jobs rely on ninja for building so those benefit as well.

The patch works by fetching the ninja clone with patches to support
jobserver-auth. Then the python script similar to the make-intercept script
calls ninja passing the input/output pipe.

Future versions of make (4.4) can pass --jobserver-auth=fifo:PATH
to pass a named pipe in which case the ninja recipe will need to be updated
to fetch the branch topic-jobserver-fifo and the intercept script will need a
minor modification.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>